Working on rendering base images...
