+++ src/app/(home)/layout.tsx
@@ -1,0 +1,12 @@
+export default function HomeLayout({
+  children, // will be a page or nested layout
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <section>
+      {/* Include shared UI here e.g. a header or sidebar */}
+      {children}
+    </section>
+  )
+}
