+++ src/app/error.tsx
@@ -1,1 +1,2 @@
 "use client";
+
