--- src/app/home/page.tsx
@@ -1,13 +1,0 @@
-import { useState, useEffect } from "react";
-import Carousel from "@/app/components/Carousel";
-
-export default function Home() {
-
-
-  return (
-    <div>
-      <h1>Home</h1>
-      <Carousel></Carousel>
-    </div>
-  )
-}
