+++ src/app/layout.tsx
@@ -6,9 +6,6 @@ import theme from '@/styles/theme'
 import { Navbar } from '@/app/components/Navbar'
 
 
-import bcrypt from 'bcrypt'
-import { createServerClient, type CookieOptions } from '@supabase/ssr'
-import { cookies } from 'next/headers'
 
 
 // const inter = Inter({ subsets: ['latin'] })
@@ -28,7 +25,7 @@ export default async function RootLayout(
     }
 ) {
 
-  const imageUrls = getImages();
+
 
   return (
     <html lang="en">
@@ -45,35 +42,3 @@ export default async function RootLayout(
     </html>
   )
 };
-
-// initial function for getting sample image set from supabase
-async function getImages(): Promise<Array<string> | undefined> {
-  const cookieStore = cookies();
-  const { SUPABASE_URL, SUPABASE_KEY } =
-    process.env;
-  const supabase = createServerClient(SUPABASE_URL!, SUPABASE_KEY!, {
-    cookies: {
-      get(name: string) {
-        return cookieStore.get(name)?.value
-      }
-    }
-  })
-  try {
-    const { data, error } = await supabase
-      .storage
-      .from('products')
-      .list('arrangements', {
-        limit: 10,
-        offset: 1
-      })
-
-    const urls = await data?.map((img): string => {
-      return img.name
-    })
-
-    return urls;
-  }
-  catch (err) {
-    console.error(err);
-  }
-}
