+++ src/app/page.tsx
@@ -1,6 +1,6 @@
-"use client"
+// "use client"
 import { useEffect, useState } from "react";
-import Home from "./home/page"
+import Home from "./(home)/page"
 
 
 export default function Main() {
